absolute path for android-core.zip
Changing the relative path to an absolute path for android-core.zip file location retrieval. NOTE: this tastes more like a hack than a fix - the root problem was that the commandline build script couldn't find any of the needed files - maybe Processing knows how to tell these paths to the build script somehow? This change needs testing on various platforms (works on Linux Mint 17 64). Tastes like a hack because I'm changing a hardcoded value to an even more hardcoded value.