File locations in commandline build script
Slightly heavy modification to allow the build script to find the required files. I would love a solution where this information is passed on to this script via either asking Processing or a settings file etc. In this version, the user has to input the locations, it's a bit meh.