Remove premultiplyMatrices and ellipseDetailMultiplier
This should simplify the code and remove some boilerplate, as discussed here https://github.com/processing/processing/issues/5256#issuecomment-366513791

I have tested this and it doesn't seem to make a significant performance difference my benchmarks. However, a further optimization could be to cache all or part of the circleDetail() calculation when tessellating a line path, since (I think?) matrices either are assumed or can be proven not to change during that time.